+++ example/lib/main.dart
@@ -55,6 +55,7 @@ class _InAppState extends State<InApp> {
       _conectionSubscription.cancel();
       _conectionSubscription = null;
     }
+    super.dispose();
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.
@@ -88,15 +89,18 @@ class _InAppState extends State<InApp> {
       print('consumeAllItems error: $err');
     }
 
-    _conectionSubscription = FlutterInappPurchase.connectionUpdated.listen((connected) {
+    _conectionSubscription =
+        FlutterInappPurchase.connectionUpdated.listen((connected) {
       print('connected: $connected');
     });
 
-    _purchaseUpdatedSubscription = FlutterInappPurchase.purchaseUpdated.listen((productItem) {
+    _purchaseUpdatedSubscription =
+        FlutterInappPurchase.purchaseUpdated.listen((productItem) {
       print('purchase-updated: $productItem');
     });
 
-    _purchaseErrorSubscription = FlutterInappPurchase.purchaseError.listen((purchaseError) {
+    _purchaseErrorSubscription =
+        FlutterInappPurchase.purchaseError.listen((purchaseError) {
       print('purchase-error: $purchaseError');
     });
   }
@@ -106,7 +110,8 @@ class _InAppState extends State<InApp> {
   }
 
   Future _getProduct() async {
-    List<IAPItem> items = await FlutterInappPurchase.instance.getProducts(_productLists);
+    List<IAPItem> items =
+        await FlutterInappPurchase.instance.getProducts(_productLists);
     for (var item in items) {
       print('${item.toString()}');
       this._items.add(item);
@@ -133,7 +138,8 @@ class _InAppState extends State<InApp> {
   }
 
   Future _getPurchaseHistory() async {
-    List<PurchasedItem> items = await FlutterInappPurchase.instance.getPurchaseHistory();
+    List<PurchasedItem> items =
+        await FlutterInappPurchase.instance.getPurchaseHistory();
     for (var item in items) {
       print('${item.toString()}');
       this._purchases.add(item);
@@ -217,8 +223,8 @@ class _InAppState extends State<InApp> {
 
   @override
   Widget build(BuildContext context) {
-    double screenWidth = MediaQuery.of(context).size.width-20;
-    double buttonWidth=(screenWidth/3)-20;
+    double screenWidth = MediaQuery.of(context).size.width - 20;
+    double buttonWidth = (screenWidth / 3) - 20;
 
     return Container(
       padding: EdgeInsets.all(10.0),