+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 5.1.0
+
+- Upgrade android billing client to `4.0.0` (#326)
+
+- Remove `orderId` in `Purchase`
+  - This is duplicate of `transactionId`.
+
 ## 5.0.4
 
 - Add iOS promo codes (#325)